Fin du composant DPE, DPE2 devient DPE
DPE était notre inspiration, dans un vieux React et un vieux style de
DPE anguleux
