+++ app/dpe/page.tsx
@@ -1,4 +1,4 @@
-import DPE from '@/components/DPE2'
+import DPE from '@/components/DPE'
 import { Intro, PageBlock } from '@/components/UI'
 import css from '@/components/css/convertToJs'
 import { Content, Wrapper } from '@/components/explications/ExplicationUI'
