+++ components/DPESelector.tsx
@@ -1,4 +1,4 @@
-import DPE from './DPE2'
+import DPE from './DPE'
 import { encodeSituation } from './publicodes/situationUtils'
 import data from '@/components/DPE.yaml'
 import css from './css/convertToJs'
